Ensure requests are authenticated for products.

+++ api/routes/products.js
@@ -21,44 +21,50 @@ router.get("/all", function(req, res) {
 });
 
 router.get("/:id", function(req, res) {
-	db.Product.findByPk(req.params.id)
-		.then( product => {
-			res.status(200).send(JSON.stringify(product));
-		})
-		.catch( err => {
-			res.status(500).send(JSON.stringify(err));
-		});
+	if (req.isAuthenticated()) {
+		db.Product.findByPk(req.params.id)
+			.then( product => {
+				res.status(200).send(JSON.stringify(product));
+			})
+			.catch( err => {
+				res.status(500).send(JSON.stringify(err));
+			});
+	}
 });
 
 router.put("/", function(req, res) {
-	db.Product.create({
-		name: req.body.name,
-		quantity: req.body.quantity,
-		expiryDate: req.body.expiryDate,
-		storageLocation: req.body.storageLocation,
-		freezable: req.body.freezable,
-		id: req.body.id
-	})
-		.then( product => {
-			res.status(200).send(JSON.stringify(product));
+	if (req.isAuthenticated()) {
+		db.Product.create({
+			name: req.body.name,
+			quantity: req.body.quantity,
+			expiryDate: req.body.expiryDate,
+			storageLocation: req.body.storageLocation,
+			freezable: req.body.freezable,
+			id: req.body.id
 		})
-		.catch( err => {
-			res.status(500).send(JSON.stringify(err));
-		});
+			.then( product => {
+				res.status(200).send(JSON.stringify(product));
+			})
+			.catch( err => {
+				res.status(500).send(JSON.stringify(err));
+			});
+	}
 });
 
 router.delete("/:id", function(req, res) {
-	db.Product.destroy({
-		where: {
-			id: req.params.id
-		}
-	})
-		.then( () => {
-			res.status(200).send();
+	if (req.isAuthenticated()){
+		db.Product.destroy({
+			where: {
+				id: req.params.id
+			}
 		})
-		.catch( err => {
-			res.status(500).send(JSON.stringify(err));
-		});
+			.then( () => {
+				res.status(200).send();
+			})
+			.catch( err => {
+				res.status(500).send(JSON.stringify(err));
+			});
+	}
 });
 
 module.exports = router;
